+++ Common/game/roomstruct.cpp
@@ -132,7 +132,7 @@ void RoomStruct::Free()
 
 void RoomStruct::FreeMessages()
 {
-    for (size_t i = 0; i < MessageCount; ++i)
+    for (uint32_t i = 0; i < MessageCount; ++i)
     {
         Messages[i].Free();
         MessageInfos[i] = MessageInfo();
@@ -145,11 +145,11 @@ void RoomStruct::FreeScripts()
     CompiledScript.reset();
 
     EventHandlers.reset();
-    for (size_t i = 0; i < HotspotCount; ++i)
+    for (uint32_t i = 0; i < HotspotCount; ++i)
         Hotspots[i].EventHandlers.reset();
     for (auto &obj : Objects)
         obj.EventHandlers.reset();
-    for (size_t i = 0; i < RegionCount; ++i)
+    for (uint32_t i = 0; i < RegionCount; ++i)
         Regions[i].EventHandlers.reset();
 }
